refactor(compaction): split merge/mod.rs by function into job/stream/file modules
